zinger: samus_pd: increment zinger hardware id minor revision
Add a minor revision to the PD device hardware ID field in the info
custom VDM and set increment this minor ID from 0 to 1 for zinger.
This differentiates zingers for the auto-update payload, so we can
update only those with a specific major and minor ID version.

BUG=none
BRANCH=samus
TEST=load onto samus and zinger. when connect zinger, see on PD
console: Dev:0x0401 SW:2289 RW:0, which shows the appropriate
device ID.

+++ common/usb_pd_protocol.c
@@ -255,7 +255,7 @@ static struct pd_protocol {
 	uint8_t vdo_count;
 
 	/* Attached ChromeOS device id & RW hash */
-	uint8_t dev_id;
+	uint16_t dev_id;
 	uint32_t dev_rw_hash[SHA1_DIGEST_SIZE/4];
 } pd[PD_PORT_COUNT];
 
@@ -1101,7 +1101,7 @@ static void pd_vdm_send_state_machine(int port)
 	}
 }
 
-static inline void pd_dev_dump_info(uint8_t dev_id, uint32_t *hash)
+static inline void pd_dev_dump_info(uint16_t dev_id, uint32_t *hash)
 {
 	int j;
 	ccprintf("Device:%d Hash:", dev_id);
@@ -1110,7 +1110,7 @@ static inline void pd_dev_dump_info(uint8_t dev_id, uint32_t *hash)
 	ccprintf("\n");
 }
 
-void pd_dev_store_rw_hash(int port, uint8_t dev_id, uint32_t *rw_hash)
+void pd_dev_store_rw_hash(int port, uint16_t dev_id, uint32_t *rw_hash)
 {
 	pd[port].dev_id = dev_id;
 	memcpy(pd[port].dev_rw_hash, rw_hash, SHA1_DIGEST_SIZE);
